修正blocklist对大文件块列表不正常的BUG.

+++ stage2/disk_io.c
@@ -31,8 +31,8 @@ unsigned long long
 block_read_func (unsigned long long buf, unsigned long long len, unsigned long write);
 
 /* instrumentation variables */
-void (*disk_read_hook) (unsigned long long, unsigned long, unsigned long) = NULL;
-void (*disk_read_func) (unsigned long long, unsigned long, unsigned long) = NULL;
+void (*disk_read_hook) (unsigned long long, unsigned long, unsigned long long) = NULL;
+void (*disk_read_func) (unsigned long long, unsigned long, unsigned long long) = NULL;
 
 /* Forward declarations.  */
 static int next_bsd_partition (void);
